Switch events_handler to listen to redis
Changes copied over from develop-2.1.0

I've tried to copy over exactly the code that has been tested in
develop, because we won't get a chance to thoroughly test it here. So
there are a few lines that do nothing (an unused import, a
self._connect where we don't do anything with the returned
connection), but I'm not removing them because this is the code that
has been tested.